+++ Platform/DirectX12/FramebufferDX1x.cpp
@@ -158,7 +158,7 @@ void Framebuffer::SetViewport(crossplatform::DeviceContext &deviceContext,float 
 	mViewport.MinDepth		= Z;
 	mViewport.MaxDepth		= D;
 
-	CD3DX12_RECT scissor(0, 0, mViewport.Width, mViewport.Height);
+	CD3DX12_RECT scissor(0, 0, (LONG)mViewport.Width, (LONG)mViewport.Height);
 
 	deviceContext.asD3D12Context()->RSSetScissorRects(1, &scissor);
 	deviceContext.asD3D12Context()->RSSetViewports(1, &mViewport);
